+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.30.1...@coveo/atomic@1.30.2) (2022-02-18)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [1.30.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.30.0...@coveo/atomic@1.30.1) (2022-02-16)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.30.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.29.5...@coveo/atomic@1.30.0) (2022-02-15)
+
+
+### Features
+
+* **atomic:** add search layout ([#1730](https://github.com/coveo/ui-kit/issues/1730)) ([aed0cae](https://github.com/coveo/ui-kit/commit/aed0caea993016ed90b941c377ccd52895146246))
+
+
+
+
+
+## [1.29.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.29.4...@coveo/atomic@1.29.5) (2022-02-11)
+
+
+### Bug Fixes
+
+* **atomic:** improved aria-live accessibility ([#1715](https://github.com/coveo/ui-kit/issues/1715)) ([124531f](https://github.com/coveo/ui-kit/commit/124531f9325e81b1e0856001f4ec4b1bc2ce8664))
+
+
+
+
+
+## [1.29.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.29.3...@coveo/atomic@1.29.4) (2022-02-09)
+
+
+### Bug Fixes
+
+* **atomic-angular:** fix custom result templates for atomic angular ([#1719](https://github.com/coveo/ui-kit/issues/1719)) ([abf0ec0](https://github.com/coveo/ui-kit/commit/abf0ec0793142a596a6660f5128d1c82a0b8063d))
+
+
+
+
+
 ## [1.29.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.29.2...@coveo/atomic@1.29.3) (2022-02-07)
 
 **Note:** Version bump only for package @coveo/atomic